upc-validator/backend/pyproject.toml
 [project]
 name = "upc-validator"
 version = "0.1.0"
 description = "UPC validation API for care labels and RFID hang tags"
 requires-python = ">=3.11"
 dependencies = [
   "fastapi>=0.110",
   "uvicorn>=0.23",
   "python-multipart>=0.0.9",
   "pymupdf>=1.23",
   "pillow>=10.0",
   "pytesseract>=0.3.10",
   "pandas>=2.1",
   "openpyxl>=3.1",
 ]
 
 [tool.uv]
 dev-dependencies = []
